lab_14_skills_multiagent_py/requirements.txt
# Lab 14 LangGraph multi-agent + skills (Python): sales proposal pipeline
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
